+++ libretro-common/include/retro_stat.h
@@ -46,7 +46,7 @@ bool path_is_character_special(const char *path);
 
 bool path_is_valid(const char *path);
 
-ssize_t path_get_size(const char *path);
+int32_t path_get_size(const char *path);
 
 /**
  * path_mkdir_norecurse:
